generate/requirements.txt
python-dotenv>=1.0.1
google-cloud-pubsub>=2.21.1
paho-mqtt>=2.1.0
pandas-gbq
